+++ conversion/clip_shape.py
