+++ models/unet_resnet101v2_pt.py
@@ -10,7 +10,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import tensorflow as tf
 
-from dataloader import DataLoader
+from loading.dataloader import DataLoader
 
 
 class Model:
